backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
pydantic==2.10.4
pydantic-settings==2.6.1
python-multipart==0.0.9
httpx==0.28.1
python-dotenv==1.0.1
google-generativeai==0.6.0
